Servo angles
Started messing with the servo angle. I remember 0.078 was the value we
got for 90 deg, but I don't remember the other ones. Go through the
OC1Interrupt and the CNInterrupt and fill in the right values for the
angles. Doing a ctrl+F and searching for OC2R should take you to the
lines that need to be filled in.