Ball Collection Complete
Finds the dispenser, reaches corner, collects 6 balls, drives to center
and aims to shoot. Code is there to start shooting, but seems that
MOSFET circuit is not working. Pin 11 goes high (tested on
oscilloscope), but motors don't turn on. Needs to be fixed before
continuing.
*Problem with frozen state: OC3 Interrupt was enabled, but flag was
never cleared. Deleting code to enable the interrupt fixed the issue.