Fine-tuning
Buttons are fixed, so went back to using button interrupt.
Right turret angle needed to be adjusted from 0.12*OC1RS to 0.125. 
Robot seems to be reaching a critical weight, so brought rotation and forward/reverse speeds down to avoid skipping which happened much too frequently.
It seems that the left IR sensor is causing problems; or, it could be the code. Probably the sensor, though, so we are going to try replacing it first.
Potential problem: If the comparator interrupt is triggered during speedup(), the robot never gets to the coasting phase and will only speedup(), stop for several seconds, proceed with slowdown(), and then start firing (only about 6 inches away from the dispenser corner). Hopefully this is only an issue while the left IR sensor is malfunctioning, because the robot should not be sensing anything until it is close to the center of the arena.